# internal/admin/handler/system.go
package handler

import (
	"github.com/gin-gonic/gin"

	"bico-admin/pkg/response"
)

// SystemHandler 系统处理器
type SystemHandler struct {
	// TODO: 添加系统服务依赖
}

// NewSystemHandler 创建系统处理器
func NewSystemHandler() *SystemHandler {
	return &SystemHandler{}
}

// GetInfo 获取系统信息
// @Summary 获取系统信息
// @Description 获取系统基本信息
// @Tags 系统管理
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/system/info [get]
func (h *SystemHandler) GetInfo(c *gin.Context) {
	// TODO: 实现获取系统信息
	response.Success(c, gin.H{
		"message": "系统信息接口待实现",
	})
}

// GetStats 获取系统统计
// @Summary 获取系统统计
// @Description 获取系统运行统计信息
// @Tags 系统管理
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/system/stats [get]
func (h *SystemHandler) GetStats(c *gin.Context) {
	// TODO: 实现获取系统统计
	response.Success(c, gin.H{
		"message": "系统统计接口待实现",
	})
}

// GetCacheStats 获取缓存统计
// @Summary 获取缓存统计
// @Description 获取缓存使用统计信息
// @Tags 系统管理
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/system/cache/stats [get]
func (h *SystemHandler) GetCacheStats(c *gin.Context) {
	// TODO: 实现获取缓存统计
	response.Success(c, gin.H{
		"message": "缓存统计接口待实现",
	})
}

// ClearCache 清理缓存
// @Summary 清理缓存
// @Description 清理系统缓存
// @Tags 系统管理
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/system/cache [delete]
func (h *SystemHandler) ClearCache(c *gin.Context) {
	// TODO: 实现清理缓存
	response.Success(c, gin.H{
		"message": "清理缓存接口待实现",
	})
}

// GetConfigList 获取配置列表
// @Summary 获取配置列表
// @Description 分页获取系统配置列表
// @Tags 配置管理
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/configs [get]
func (h *SystemHandler) GetConfigList(c *gin.Context) {
	// TODO: 实现获取配置列表
	response.Success(c, gin.H{
		"message": "配置列表接口待实现",
	})
}

// GetConfig 获取配置
// @Summary 获取配置详情
// @Description 根据ID获取配置详情
// @Tags 配置管理
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "配置ID"
// @Success 200 {object} response.ApiResponse
// @Router /admin/configs/{id} [get]
func (h *SystemHandler) GetConfig(c *gin.Context) {
	// TODO: 实现获取配置
	response.Success(c, gin.H{
		"message": "获取配置接口待实现",
	})
}

// CreateConfig 创建配置
// @Summary 创建配置
// @Description 创建新的系统配置
// @Tags 配置管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/configs [post]
func (h *SystemHandler) CreateConfig(c *gin.Context) {
	// TODO: 实现创建配置
	response.Success(c, gin.H{
		"message": "创建配置接口待实现",
	})
}

// UpdateConfig 更新配置
// @Summary 更新配置
// @Description 更新系统配置
// @Tags 配置管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "配置ID"
// @Success 200 {object} response.ApiResponse
// @Router /admin/configs/{id} [put]
func (h *SystemHandler) UpdateConfig(c *gin.Context) {
	// TODO: 实现更新配置
	response.Success(c, gin.H{
		"message": "更新配置接口待实现",
	})
}

// DeleteConfig 删除配置
// @Summary 删除配置
// @Description 删除系统配置
// @Tags 配置管理
// @Security ApiKeyAuth
// @Param id path int true "配置ID"
// @Success 200 {object} response.ApiResponse
// @Router /admin/configs/{id} [delete]
func (h *SystemHandler) DeleteConfig(c *gin.Context) {
	// TODO: 实现删除配置
	response.Success(c, gin.H{
		"message": "删除配置接口待实现",
	})
}

// GetLogList 获取日志列表
// @Summary 获取日志列表
// @Description 分页获取系统日志列表
// @Tags 日志管理
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/logs [get]
func (h *SystemHandler) GetLogList(c *gin.Context) {
	// TODO: 实现获取日志列表
	response.Success(c, gin.H{
		"message": "日志列表接口待实现",
	})
}

// ClearLogs 清理日志
// @Summary 清理日志
// @Description 清理系统日志
// @Tags 日志管理
// @Security ApiKeyAuth
// @Success 200 {object} response.ApiResponse
// @Router /admin/logs [delete]
func (h *SystemHandler) ClearLogs(c *gin.Context) {
	// TODO: 实现清理日志
	response.Success(c, gin.H{
		"message": "清理日志接口待实现",
	})
}